+++ CHANGELOG.md
@@ -39,6 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- [#1455](https://github.com/alleslabs/celatone-frontend/pull/1455) Update transactions table tx decoder column
+- [#1454](https://github.com/alleslabs/celatone-frontend/pull/1454) Update react-app-shell to v1.3.0
 - [#1452](https://github.com/alleslabs/celatone-frontend/pull/1452) Support non-initia chain config
 - [#1451](https://github.com/alleslabs/celatone-frontend/pull/1451) Update lowest threshold decimal points
 - [#1444](https://github.com/alleslabs/celatone-frontend/pull/1444) Sort NFTs and NFT collections by ASC order
@@ -54,6 +56,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#1456](https://github.com/alleslabs/celatone-frontend/pull/1456) Upgrade Tx Decoder to 0.7.1
 - [#1453](https://github.com/alleslabs/celatone-frontend/pull/1453) Upgrade InterwovenKit to 2.0.0
 - [#1450](https://github.com/alleslabs/celatone-frontend/pull/1450) Remove extra query key in useNftsSequencer
 - [#1446](https://github.com/alleslabs/celatone-frontend/pull/1446) Update background main color for Initia theme and default theme